functions/app/app/requirements.txt
google-cloud-pubsub==0.39.1
requests==2.21.0